+++ packages/graphyne-ws/package.json
@@ -34,6 +34,7 @@
     "ws": "^7.0.0"
   },
   "devDependencies": {
+    "@types/node": "^14.0.23",
     "@types/ws": "^7.2.6"
   },
   "engines": {
